infra/requirements.txt
backports.ssl-match-hostname==3.5.0.1
boto3==1.9.19
botocore==1.12.19
certifi==2018.8.24
chardet==3.0.4
coloredlogs==10.0
docker==3.5.0
docker-pycreds==0.3.0
docutils==0.14
executor==21.2
fasteners==0.14.1
futures==3.2.0
gitdb2==2.0.4
GitPython==2.1.11
humanfriendly==4.17
idna==2.7
ipaddress==1.0.22
jmespath==0.9.3
monotonic==1.5
property-manager==2.3.1
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
smmap2==2.0.4
urllib3==1.23
verboselogs==1.7
websocket-client==0.53.0
